#  Simplify the Parser monad implementation.