--- a/proto-lens/src/Data/ProtoLens/Encoding/Parser/Internal.hs
+++ b/proto-lens/src/Data/ProtoLens/Encoding/Parser/Internal.hs
@@ -1,12 +1,12 @@
+{-# LANGUAGE LambdaCase #-}
 -- | Definition of the parsing monad, plus internal
 -- unsafe functions.
 module Data.ProtoLens.Encoding.Parser.Internal
     ( Parser(..)
-    , ParserResult(..)
+    , ParseResult(..)
     ) where
 
 import Control.Monad (ap)
-import Control.Monad.Trans.Except
 import Data.Word (Word8)
 import Foreign.Ptr
 
@@ -14,27 +14,30 @@ import Foreign.Ptr
 newtype Parser a = Parser
     { unParser :: Ptr Word8 -- End position of the input
                -> Ptr Word8 -- Current position in the input
-               -> ExceptT String IO (ParserResult a)
+               -> IO (ParseResult a)
     }
 
-data ParserResult a = ParserResult
-    { _newPos :: !(Ptr Word8) -- ^ New position in the input
-    , unParserResult :: a
-    }
+data ParseResult a
+    = ParseSuccess
+        { _newPos :: !(Ptr Word8) -- ^ New position in the input
+        , unParserResult :: a
+        }
+    | ParseFailure String
 
-instance Functor ParserResult where
-    fmap f (ParserResult p x) = ParserResult p (f x)
+instance Functor ParseResult where
+    fmap f (ParseSuccess p x) = ParseSuccess p (f x)
+    fmap _ (ParseFailure s) = ParseFailure s
 
 instance Functor Parser where
     fmap f (Parser g) = Parser $ \end cur -> fmap f <$> g end cur
 
 instance Applicative Parser where
-    pure x = Parser $ \_ cur -> return $ ParserResult cur x
+    pure x = Parser $ \_ cur -> return $ ParseSuccess cur x
     (<*>) = ap
 
 instance Monad Parser where
-    fail s = Parser $ \_ _ -> throwE s
+    fail s = Parser $ \_ _ -> return $ ParseFailure s
     return = pure
-    Parser f >>= g = Parser $ \end pos -> do
-        ParserResult pos' x <- f end pos
-        unParser (g x) end pos'
+    Parser f >>= g = Parser $ \end pos -> f end pos >>= \case
+        ParseSuccess pos' x -> unParser (g x) end pos'
+        ParseFailure s -> return $ ParseFailure s